Moved order test from test-orderRoute to test-orderModel
