+++ test/test-orderRoute.js
@@ -1,6 +1,5 @@
 import request from 'request';
 import { expect, assert } from 'chai';
-import { Order } from '../models/orderModel';
 
 
 describe('Api v1 Server', () => {
@@ -26,34 +25,6 @@ describe('Api v1 Server', () => {
       assert(typeof (data.body), 'json');
     });
   });
-
-  describe('initialise orderManager with no data', () => {
-    let orderManager;
-    before((done) => {
-      orderManager = new Order();
-      done();
-    });
-    it('has empty order', () => {
-      expect(JSON.stringify(orderManager.orders)).to.equal(JSON.stringify({}));
-    });
-
-    it('has lastOrderId 0', () => {
-      expect(orderManager.lastOrderId).to.equal(0);
-    });
-  });
-
-  describe('try to call order class as function', () => {
-    it('throws error', () => {
-      try {
-        const orderManager = Order();
-        expect(typeof (orderManager)).to.equal('undefined');
-      } catch (err) {
-        expect(err.message).to.equal('Cannot call a class as a function');
-        expect(typeof (orderManager)).to.equal('undefined');
-      }
-    });
-  });
-
 
   describe('POST /orders - form', () => {
     const data = {};
